🔨 Large internal refactor / file management
